+++ src/services/blood_spawner.rs
@@ -5,7 +5,7 @@ pub struct BloodSpawnerRequest {
     pub y: i32,
     pub fg: RGB,
     pub bg: RGB,
-    pub glyph: u8,
+    pub glyph: u16,
     pub level: i32,
 }
 
@@ -20,7 +20,7 @@ impl BloodSpawner {
         }
     }
 
-    pub fn request(&mut self, x: i32, y: i32, fg: RGB, bg: RGB, glyph: u8, level: i32) {
+    pub fn request(&mut self, x: i32, y: i32, fg: RGB, bg: RGB, glyph: u16, level: i32) {
         self.requests.push(BloodSpawnerRequest {
             x,
             y,
