+++ src/services/particle_effect_spawner.rs
@@ -5,7 +5,7 @@ pub struct ParticleEffectSpawnerRequest {
     pub y: i32,
     pub fg: RGB,
     pub bg: RGB,
-    pub glyph: u8,
+    pub glyph: u16,
     pub lifetime: f32,
     pub level: i32,
 }
@@ -27,7 +27,7 @@ impl ParticleEffectSpawner {
         y: i32,
         fg: RGB,
         bg: RGB,
-        glyph: u8,
+        glyph: u16,
         lifetime: f32,
         level: i32,
     ) {
